style(AddRadioComponentForm): класс RadioComponentCreatedEventArgs перенесен в файл RadioComponentCreatedEventArgs.cs
