+++ Lab1View/Lab1View.csproj
@@ -61,6 +61,7 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RadioComponentCreatedEventArgs.cs" />
     <Compile Include="SearchRadioComponentForm.cs">
       <SubType>Form</SubType>
     </Compile>
