+++ Lab1View/RadioComponentCreatedEventArgs.cs
@@ -1,0 +1,31 @@
+﻿#define TEST
+
+using System;
+using Lab1Model;
+
+
+namespace Lab1View
+{
+	/// <summary>
+	/// Класс данных события создания нового радиокомпонента
+	/// </summary>
+	public class RadioComponentCreatedEventArgs : EventArgs
+    {
+        /// <summary>
+        /// Создает объект класса
+        /// <see cref="RadioComponentCreatedEventArgs"/>
+        /// </summary>
+        /// <param name="radioComponent">
+        /// Созданный объект радиокомпонента</param>
+        public RadioComponentCreatedEventArgs(
+            RadioComponentBase radioComponent)
+        {
+            RadioComponent = radioComponent;
+        }
+
+        /// <summary>
+        /// Объект радиокомпонента
+        /// </summary>
+        public RadioComponentBase RadioComponent { get; }
+    }
+}
